+++ dockerfile_builder/src/instruction_builder.rs
@@ -1057,285 +1057,3 @@ impl ShellBuilder {
         Ok(format!(r#"["{}"{}]"#, self.executable, params))
     }
 }
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-    use expect_test::expect;
-
-    #[test]
-    fn from() {
-        let from = FromBuilder::builder().image("cargo-chef").build().unwrap();
-        let expected = expect!["FROM cargo-chef"];
-        expected.assert_eq(&from.to_string());
-
-        let from = FromBuilder::builder()
-            .image("cargo-chef")
-            .platform("linux/arm64")
-            .build()
-            .unwrap();
-        let expected = expect!["FROM --platform=linux/arm64 cargo-chef"];
-        expected.assert_eq(&from.to_string());
-
-        let from = FromBuilder::builder()
-            .image("cargo-chef")
-            .name("chef")
-            .tag("latest")
-            .build()
-            .unwrap();
-        let expected = expect!["FROM cargo-chef:latest AS chef"];
-        expected.assert_eq(&from.to_string());
-
-        let from = FromBuilder::builder()
-            .image("cargo-chef")
-            .name("chef")
-            .digest("sha256")
-            .build()
-            .unwrap();
-        let expected = expect!["FROM cargo-chef@sha256 AS chef"];
-        expected.assert_eq(&from.to_string());
-    }
-
-    #[test]
-    fn run() {
-        let run = RunExecBuilder::builder()
-            .executable("mybin.exe")
-            .build()
-            .unwrap();
-        let expected = expect![r#"RUN ["mybin.exe"]"#];
-        expected.assert_eq(&run.to_string());
-    }
-
-    #[test]
-    fn cmd() {
-        let cmd = CmdBuilder::builder()
-            .command("some command")
-            .build()
-            .unwrap();
-        let expected = expect!["CMD some command"];
-        expected.assert_eq(&cmd.to_string());
-
-        let cmd = CmdExecBuilder::builder()
-            .executable("some command")
-            .build()
-            .unwrap();
-        let expected = expect![r#"CMD ["some command"]"#];
-        expected.assert_eq(&cmd.to_string());
-    }
-
-    #[test]
-    fn from_err() {
-        let from = FromBuilder::builder()
-            .image("cargo-chef")
-            .tag("t")
-            .digest("d")
-            .build();
-        match from {
-            Ok(_) => panic!("Both tag and digest are set. Expect test to fail"),
-            Err(e) => assert_eq!(
-                e.to_string(),
-                "Dockerfile image can only have tag OR digest".to_string(),
-            ),
-        }
-    }
-
-    #[test]
-    fn expose() {
-        let expose = ExposeBuilder::builder().port(80).build().unwrap();
-        let expected = expect!["EXPOSE 80"];
-        expected.assert_eq(&expose.to_string());
-    }
-
-    #[test]
-    fn add() {
-        let add = AddBuilder::builder()
-            .src("hom*")
-            .dest("/mydir/")
-            .build()
-            .unwrap();
-        let expected = expect!["ADD hom* /mydir/"];
-        expected.assert_eq(&add.to_string());
-    }
-
-    #[test]
-    fn add_http() {
-        let add = AddHttpBuilder::builder()
-            .checksum("sha256::123")
-            .src("http://example.com/foobar")
-            .dest("/")
-            .build()
-            .unwrap();
-        let expected = expect!["ADD --checksum=sha256::123 http://example.com/foobar /"];
-        expected.assert_eq(&add.to_string());
-    }
-
-    #[test]
-    fn add_git() {
-        let add = AddGitBuilder::builder()
-            .keep_git_dir(true)
-            .git_ref("https://github.com/moby/buildkit.git#v0.10.1")
-            .dir("/buildkit")
-            .build()
-            .unwrap();
-        let expected = expect![
-            "ADD --keep-git-dir=true https://github.com/moby/buildkit.git#v0.10.1 /buildkit"
-        ];
-        expected.assert_eq(&add.to_string());
-    }
-
-    #[test]
-    fn copy() {
-        let copy = CopyBuilder::builder()
-            .link(true)
-            .src("foo/")
-            .dest("bar/")
-            .build()
-            .unwrap();
-        let expected = expect!["COPY --link foo/ bar/"];
-        expected.assert_eq(&copy.to_string());
-    }
-
-    #[test]
-    fn entrypoint() {
-        let entrypoint = EntrypointBuilder::builder()
-            .command("some command")
-            .build()
-            .unwrap();
-        let expected = expect![r#"ENTRYPOINT some command"#];
-        expected.assert_eq(&entrypoint.to_string());
-
-        let entrypoint = EntrypointExecBuilder::builder()
-            .executable("mybin.exe")
-            .build()
-            .unwrap();
-        let expected = expect![r#"ENTRYPOINT ["mybin.exe"]"#];
-        expected.assert_eq(&entrypoint.to_string());
-
-        let entrypoint = EntrypointExecBuilder::builder()
-            .executable("top")
-            .param("-b")
-            .build()
-            .unwrap();
-        let expected = expect![r#"ENTRYPOINT ["top", "-b"]"#];
-        expected.assert_eq(&entrypoint.to_string());
-    }
-
-    #[test]
-    fn volume() {
-        let volume = VolumeBuilder::builder()
-            .path("/myvol1")
-            .path("/myvol2")
-            .build()
-            .unwrap();
-        let expected = expect!["VOLUME /myvol1 /myvol2"];
-        expected.assert_eq(&volume.to_string());
-    }
-
-    #[test]
-    fn user() {
-        let user = UserBuilder::builder().user("myuser").build().unwrap();
-        let expected = expect!["USER myuser"];
-        expected.assert_eq(&user.to_string());
-
-        let user = UserBuilder::builder()
-            .user("myuser")
-            .group("mygroup")
-            .build()
-            .unwrap();
-        let expected = expect!["USER myuser:mygroup"];
-        expected.assert_eq(&user.to_string());
-    }
-
-    #[test]
-    fn workdir() {
-        let workdir = WorkdirBuilder::builder()
-            .path("/path/to/workdir")
-            .build()
-            .unwrap();
-        let expected = expect!["WORKDIR /path/to/workdir"];
-        expected.assert_eq(&workdir.to_string());
-    }
-
-    #[test]
-    fn arg() {
-        let arg = ArgBuilder::builder().name("user1").build().unwrap();
-        let expected = expect!["ARG user1"];
-        expected.assert_eq(&arg.to_string());
-
-        let arg = ArgBuilder::builder()
-            .name("user1")
-            .value("someuser")
-            .build()
-            .unwrap();
-        let expected = expect!["ARG user1=someuser"];
-        expected.assert_eq(&arg.to_string());
-    }
-
-    #[test]
-    fn onbuild() {
-        let onbuild = OnbuildBuilder::builder()
-            .instruction(Instruction::ADD(ADD::from(". /app/src")))
-            .build()
-            .unwrap();
-        let expected = expect!["ONBUILD ADD . /app/src"];
-        expected.assert_eq(&onbuild.to_string());
-    }
-
-    #[test]
-    fn onbuild_err() {
-        let onbuild = OnbuildBuilder::builder()
-            .instruction(Instruction::ONBUILD(ONBUILD::from("RUN somecommand")))
-            .build();
-        match onbuild {
-            Ok(_) => panic!("Chaining Onbuild instructions. Expect test to fail"),
-            Err(e) => assert_eq!(
-                e.to_string(),
-                "Chaining ONBUILD instructions using ONBUILD ONBUILD isn’t allowed".to_string(),
-            ),
-        }
-
-        let onbuild = OnbuildBuilder::builder()
-            .instruction(Instruction::FROM(FROM::from("someimage")))
-            .build();
-        match onbuild {
-            Ok(_) => {
-                panic!("ONBUILD instruction may not trigger FROM instruction. Expect test to fail")
-            }
-            Err(e) => assert_eq!(
-                e.to_string(),
-                "ONBUILD instruction may not trigger FROM instruction".to_string(),
-            ),
-        }
-    }
-
-    #[test]
-    fn stopsignal() {
-        let stopsignal = StopsignalBuilder::builder()
-            .signal("SIGKILL")
-            .build()
-            .unwrap();
-        let expected = expect!["STOPSIGNAL SIGKILL"];
-        expected.assert_eq(&stopsignal.to_string());
-    }
-
-    #[test]
-    fn healthcheck() {
-        let healthcheck = HealthcheckBuilder::builder()
-            .cmd(CMD::from("curl -f http://localhost/"))
-            .build()
-            .unwrap();
-        let expected = expect!["HEALTHCHECK CMD curl -f http://localhost/"];
-        expected.assert_eq(&healthcheck.to_string());
-
-        let healthcheck = HealthcheckBuilder::builder()
-            .cmd(CMD::from("curl -f http://localhost/"))
-            .interval(15)
-            .timeout(200)
-            .start_period(5)
-            .retries(5)
-            .build()
-            .unwrap();
-        let expected = expect!["HEALTHCHECK --interal=15 --timeout=200 --start-period=5 --retries=5 CMD curl -f http://localhost/"];
-        expected.assert_eq(&healthcheck.to_string());
-    }
-}
